+++ OP_camera_controls.py
@@ -3,6 +3,7 @@ from bpy.types import Operator
 from mathutils import Vector
 
 
+''' ## ops based version, only valid in object mode
 class STORYTOOLS_OT_camera_pan(Operator):
     bl_idname = "storytools.camera_pan"
     bl_label = 'Pan Camera'
@@ -24,26 +25,101 @@ class STORYTOOLS_OT_camera_pan(Operator):
         return self.execute(context)
 
     def execute(self, context):
+        ## !!! Work only in object mode
         orientation = 'VIEW'
-        ## ? switch to local outside of cam ? seem not needed
-        # if context.space_data.region_3d.view_perspective != 'CAMERA':
-        #     orientation = 'LOCAL'
-
         with context.temp_override(selected_objects=[context.scene.camera]):
             bpy.ops.transform.translate('INVOKE_DEFAULT',
-                orient_type=orientation, constraint_axis=self.constraint_axis)
+                orient_type=orientation, constraint_axis=self.constraint_axis)        
+        return {"FINISHED"}
+'''
+
+class STORYTOOLS_OT_camera_pan(Operator):
+    bl_idname = "storytools.camera_pan"
+    bl_label = 'Object Pan Translate'
+    bl_description = "Translate active object"
+    bl_options = {'REGISTER', 'UNDO', 'INTERNAL'}
+
+    @classmethod
+    def poll(cls, context):
+        return context.scene.camera
+
+    def invoke(self, context, event):
+        self.cam = context.scene.camera
+        # self.use_x, self.use_y = True
+        self.lock = None
+        self.init_pos = self.cam.location.copy() # to restore if cancelled
+        self.init_mouse = Vector((event.mouse_x, event.mouse_y))
+        self.lock_text = 'Camera Pan'
+        self.local_x = self.cam.matrix_world.to_quaternion() @ Vector((1,0,0))
+        self.local_y = self.cam.matrix_world.to_quaternion() @ Vector((0,1,0))
+        context.window.cursor_set("SCROLL_XY")
+        context.window_manager.modal_handler_add(self)
+        self.update_position(context, event)
+        return {'RUNNING_MODAL'}
+
+    def update_position(self, context, event):
+        mouse_co = Vector((event.mouse_x, event.mouse_y))
+        lock = self.lock
         
-        # transform.translate(value=(0, 0, 0), orient_axis_ortho='X', orient_type='GLOBAL', 
-        # orient_matrix=((0, 0, 0), (0, 0, 0), (0, 0, 0)), orient_matrix_type='GLOBAL', 
-        # constraint_axis=(False, False, False), mirror=False, use_proportional_edit=False, 
-        # proportional_edit_falloff='SMOOTH', proportional_size=1, use_proportional_connected=False, 
-        # use_proportional_projected=False, snap=False, snap_elements={'INCREMENT'}, use_snap_project=False, 
-        # snap_target='CLOSEST', use_snap_self=True, use_snap_edit=True, use_snap_nonedit=True, 
-        # use_snap_selectable=False, snap_point=(0, 0, 0), snap_align=False, snap_normal=(0, 0, 0), 
-        # gpencil_strokes=False, cursor_transform=False, texture_space=False, remove_on_cancel=False, 
-        # view2d_edge_pan=False, release_confirm=False, use_accurate=False, use_automerge_and_split=False)
+        ## Slower with shift
+        fac = 0.01 if event.shift else 0.1
+
+        move_2d = mouse_co - self.init_mouse
         
-        ## for more control need a custom pan modal
+        # Ctrl: override lock to "major" direction
+        if event.ctrl:
+            if abs(move_2d.x) >= abs(move_2d.y):
+                lock = 'X'
+            else:
+                lock = 'Y'
+
+        move_vec = Vector((0,0,0))
+        if not lock or lock == 'X': 
+            move_vec += self.local_x * (move_2d.x * fac)
+        if not lock or lock == 'Y': 
+            move_vec += self.local_y * (move_2d.y * fac)
+
+        # set location
+        self.cam.location = self.init_pos + move_vec
+        
+        ## set header text (optional)
+        local_move = move_2d * fac
+        self.lock_text = f'Camera Pan X: {local_move.x:.3f}, Y: {local_move.y:.3f}'
+        self.lock_text += f' | Lock Axis {lock}' if lock else ''
+        context.area.header_text_set(self.lock_text)
+
+    def modal(self, context, event):
+        self.update_position(context, event)
+        
+        if event.type in ('X','Y') and event.value == 'PRESS':
+            self.lock = event.type if self.lock != event.type else None
+        
+        elif event.type == 'LEFTMOUSE': # and event.value == 'RELEASE'
+            context.window.cursor_set("DEFAULT")
+            # set key autokeying
+            if context.scene.tool_settings.use_keyframe_insert_auto:
+                self.cam.keyframe_insert('location')
+                # Better to insert all
+                self.cam.keyframe_insert('rotation_euler')
+                self.cam.keyframe_insert('scale')
+
+            self.stop(context)
+            return {'FINISHED'}
+
+        elif event.type in {'RIGHTMOUSE', 'ESC'}:
+            self.cam.location = self.init_pos
+            self.stop(context)
+            return {'CANCELLED'}
+        
+        return {'RUNNING_MODAL'}
+    
+    def stop(self, context):
+        # remove draw handler and text set
+        context.area.header_text_set(None) # reset header
+        context.window.cursor_set("DEFAULT")
+        context.area.tag_redraw()
+    
+    def execute(self, context):
         return {"FINISHED"}
 
 # Not really needed, already in Grease pencil tools