+++ constants.py
@@ -1,0 +1,1 @@
+LAYERMAT_PREFIX = 'mat--'